refactored the files as the slides told us to do - ModuleApplication is part of package "mops" - changed the Controller-Path to ("/module")
