+++ src/main/java/mops/ModuleApplication.java
@@ -1,4 +1,4 @@
-package mops.module;
+package mops;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
