+++ src/main/java/mops/module/indexController.java
@@ -1,10 +1,11 @@
 package mops.module;
 
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
+@RequestMapping("/module")
 public class indexController {
 
     @GetMapping("/")
